+++ packages/analytics-plugin-trench/CHANGELOG.md
@@ -1,5 +1,17 @@
 # analytics-plugin-trench
 
+## 0.0.9
+
+### Patch Changes
+
+- f286470: Enables batching as well as click tracking in autocapture mode
+
+## 0.0.8
+
+### Patch Changes
+
+- 66e98aa: Fixes an issue where `identify` is not properly setting the traits
+
 ## 0.0.7
 
 ### Patch Changes